app/requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
kafka-python
faker
drf-yasg[validation]
drf-spectacular
drf-spectacular-sidecar
pytest
pytest-django